cleaned up code; deleted main.rb & renamed main_restructured.rb to main.rb
